pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weekly-report"
version = "0.1.0"
description = "Weekly report PDF pipeline with CSV validation and deterministic outputs"
authors = [{name = "CDLP", email = "dev@cdlp.se"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
    "numpy>=1.24.4,<2.0",
    "pandera>=0.17.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "loguru>=0.7.0",
    "plotly>=6.1.1",
    "kaleido>=0.2.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "reportlab>=4.0.0",
    "fastapi>=0.104.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.24.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
weekly-report = "weekly_report.src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["weekly_report*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
